Remove store_part[i]==-1 condition (checked)
So in the `phase_comm_table()` function, which fills the comm table
for a given phase, I removed the store_part[i_vvidx] == -1 condition,
because it caused trouble when developing the new prep function.

But I did check it with an assert, and it seems to have been
redundant.

Emil