[BIG REFACTOR] introduced `iterator`
Instead of `fill_mcounts()`, `fill_idx_mbufs()`,
`skirt_fill_mcounts()` and `skirt_fill_idx_mbufs()` there is now a
single `iterator()` function, which fills only the counts if the first
`fill` parameter is 0 and and fills the index mbuf otherwise.  If the
second parameter is `skirt_cond` then is acts as a "skir" function, if
as a non-skirt function if it is `phase_cond`.