Add `cur_part` to `partial_cd` class
Since at multiple points of code, the partition id of the processed
vertex was needed, it seemed logical to store it in a new variable
instead always looking it up by index (both index passing and looking
up in the `partitions` array might lead to inefficiency).

For this reason `cur_part` is set at the beginning of `proc_vertex`.
It is a little bit "ugly" in the sense of having a "global" variable
in the class, but hopefully it brings more joy than pain.