Replace `get_store_part()` with `at()` method call
Since `set_store_part()` was renamed, `get_store_part()` seemed "odd"
to be all alone.  Also the `at()` method did basically what you need:
get an element + a check (just in case, sort of an assertion). And it
is used only once, so I'm pretty sure it was a good idea to replace
it.