Add half finished `init_idcs` insertation
If we are at a "dirty" vertex, i.e. it is partially computed, we need
to copy the partially computed part from the previous phase.  There
are two options, the partially computed vertex was on a different
partition in the previous phase, or it was on the same partition.  In
the former case it needs to be put into the `init_dics` and
`proc_init_dict` takes care of the rest (not yet implemented), in the
later case it can (and should) be inserted directly into the
`init_idcs`.