Add MPI support
`npart` is now read from MPI_Comm_size instead of `argv[2]`.

This is still very sub-optimal, i.e. we all the preparations on each
MPI process, but the main thing is the exec part: that is where we'll
be able to call the MPI_Alltoallv().

The debug call (which prints all the nice levels and such) is also
removed, because MPI just messes it up by outputting the same thing 4x
from each process.