Rearrange `partial_cd` constructor
So, the order here is important. The first call,
i.e. `metis_partition` should be ok, since it is independent of
everything (this is something like the finalisation of phase == -1).
The second call, `init_communication()` is basically filling the
elements of the array - it should be there, because it deals with the
data which is "already there" (before anything happens).  Then stuff
which happens at the beginning of each phase is executed.