Add `result_idx`
Although `mbuf` holds all the results (including the intermediate
ones), it is not too helpful on it's own, because it holds some
partial results as well.  Although at this point we expect the value
of 1.0 everywhere, the partial vertices with values <1.0 will also be
stored in `mbuf` and it is not going to be obvious if the results are
correct or not: it will not be clear if the values <1.0 are wrong
results or partial results (which are ok).