Add vector index to `result_idx`
One of the "quick and dirty" solutions was to expect all 1.0 values as
the output but now that is changing and this is the first step towards
a more sophisticated approach.

Instead of keeping the values of the resulting vector in unspecified
order, now `result_idx` keeps both the index where a element can be
found in the `mbuf` and the index where it belongs to in the resulting
vector, i.e. all the information needed to reconstruct the result.