Change `vector` to `set` in comm. dicts
It came up during debugging that some of the communication
dictionaries (`comm_dict` and/or `init_dict`) might have duplicate
entries (which is obviously redundant).  For this reason, the
container of the "things that need to be transmitted" was changed from
a `std::vector` to a `std::set` to eliminate duplicate entries. This
required the following (additional) changes:

- The `recvcounts` and `sendcounts` can not be incremented by 1 any
  more, they have to be calculated from the set sizes.  This is done
  by applying the new `UpdateMPICountBuffers()` function to all
  elements of the `comm_dict` and `init_dict`.

- The `push_back` methods of vectors needed to be replaced with the
  `insert` method of sets.

- The `ProcCommDict` and `ProcInitDict` methods needed to be updated
  accordingly.