Change `comm_dict` to skip redundant vertices
For multiple `kMcol` type communications, the same vertex was
transmitted twice.  Having a set of `(src_idx, tgt_idx, type)` tuples
was not enough, since some times for the same `src_idx` the `tgt_idx`
is different.  To fix this, now `comm_dict` maps from `(src_idx,
type)` pairs to a map.  The map `comm_dict[{src_idx, type}]` is the
set of all `tgt_idx` indices which need to be updated in
`ProcCommDict`.