Add `Results::Filename()` => parallel testing
Running test on the same matrix with different parameters (different
npart and nlevel) values, produced intermediate and result files with
the same name.  As a result, running the different tests, overwrote
each others files.  With the last commit (which provided the
`Filename()` method for the `Buffers` class) this addition enables all
tests to write files with unique names, which in turn enables testing
to be run in parallel using:

  ctest -j $(nproc)